[#13] refactor: 반찬 response와 반찬 모델 파일 분리
